Fix stale network sequence handling
Increases packet sequence IDs to uint32_t and tracks the latest sequence/transaction per session so stale multiplayer updates are ignored. It also accepts the first StartGamePacket during reconnects and resets transaction state when reconnecting, preventing old game state from corrupting the lobby or team shop.

+++ cpp/NetworkProvider.h
@@ -22,7 +22,7 @@ struct PlayerStatePacket {
     uint8_t  type      = 1;
     uint8_t  protocolVersion = 1;
     uint32_t playerId;
-    uint16_t sequenceId; // Added for ordering
+    uint32_t sequenceId; // Added for ordering (was uint16_t, now uint32_t to prevent wrap)
     float    x;
     float    y;
     float    vx;
